refactor: move browse, tabs, terminal panel into modules — god trio now thin
render.rs 6411→~920 lines, app.rs ~3450→~757. main/app/render now hold only
the render dispatch, controller core (new/repo/refresh/menus/save), and shared
shell helpers (with_scrollbars, editor_island_w, render_context_menu); every
feature lives in its own src/<feature>.rs.
