+++ src/app.rs
@@ -552,243 +552,14 @@ impl Kyde {
 
 
 
-    /// Expand/collapse a directory in the Browse tree.
-    pub(crate) fn toggle_dir(&mut self, dir: PathBuf, cx: &mut Context<Self>) {
-        if !self.expanded.remove(&dir) {
-            self.expanded.insert(dir);
-        }
-        cx.notify();
-    }
 
-    /// "Select Opened File in Tree" (IntelliJ-style): switch to Browse, expand
-    /// every ancestor of the active file, select its row, and scroll it into
-    /// view. Falls back to the highlighted row if no file is open.
-    pub(crate) fn reveal_in_tree(&mut self, cx: &mut Context<Self>) {
-        let Some(target) = self
-            .open_path
-            .clone()
-            .or_else(|| self.selected_path.clone())
-        else {
-            return;
-        };
-        self.mode = Mode::Browse;
-        // Expand every ancestor dir (incl. the root `""`) so the row is visible.
-        for anc in target.ancestors().skip(1) {
-            self.expanded.insert(anc.to_path_buf());
-        }
-        self.selected_path = Some(target.clone());
-        // Find the target's index in the same flattened order render_browse uses
-        // (root row, then tree rows, then scratches) and scroll it into view.
-        let mut idx = if target.as_os_str().is_empty() {
-            Some(0)
-        } else {
-            None
-        };
-        if idx.is_none() {
-            let mut i = 1usize;
-            for r in self.file_tree.visible(&self.expanded) {
-                if r.path == target {
-                    idx = Some(i);
-                    break;
-                }
-                i += 1;
-            }
-            if idx.is_none() {
-                for s in &self.scratches {
-                    if *s == target {
-                        idx = Some(i);
-                        break;
-                    }
-                    i += 1;
-                }
-            }
-        }
-        if let Some(i) = idx {
-            self.tree_scroll.scroll_to_item(i);
-        }
-        cx.notify();
-    }
 
-    /// Open `rel` as a permanent tab (double-click, finder, "Open", restore, …). If it was
-    /// the preview tab, it's promoted (the preview slot clears).
-    pub(crate) fn open_file(&mut self, rel: PathBuf, cx: &mut Context<Self>) {
-        self.open_file_inner(rel, false, cx);
-    }
 
-    /// Open `rel` as the VS Code-style *preview* (temporary) tab: a single-click in the tree.
-    /// Reuses the one preview slot — a subsequent single-click on a different file replaces it
-    /// in place rather than opening a new tab. Clicking a file that's already open as a
-    /// permanent tab just activates it (no demotion).
-    pub(crate) fn preview_file(&mut self, rel: PathBuf, cx: &mut Context<Self>) {
-        self.open_file_inner(rel, true, cx);
-    }
 
-    fn open_file_inner(&mut self, rel: PathBuf, preview: bool, cx: &mut Context<Self>) {
-        // Images preview via `img()` and font files preview in their own typeface (see
-        // render_browse) — don't load their binary bytes into the text editor.
-        if !is_image(&rel) && !is_font_file(&rel) {
-            // Scratch files live outside the repo (absolute paths) — read them straight
-            // from disk. Repo-relative files go through the repo's working tree when this is
-            // a git repo, else straight from disk under the project root (non-git Browse).
-            let content = if rel.is_absolute() {
-                std::fs::read_to_string(&rel).unwrap_or_default()
-            } else if let Some(repo) = self.repo() {
-                repo.working_content(&rel).ok().unwrap_or_default()
-            } else if let Some(root) = self.repo_root.as_ref() {
-                std::fs::read_to_string(root.join(&rel)).unwrap_or_default()
-            } else {
-                String::new()
-            };
-            let lang = self.effective_lang(&rel);
-            self.file_editor.update(cx, |e, cx| {
-                e.line_numbers = true;
-                e.set_content(content, lang, cx);
-            });
-            // Point the editor at whichever scroll container it renders in, so caret-follow
-            // and drag auto-scroll move the right one: the Markdown split uses
-            // `md_editor_scroll`, plain Browse uses `file_scroll` (mirrors the `md` gate in
-            // render_browse).
-            let md = matches!(highlight::Lang::from_path(&rel), highlight::Lang::Markdown)
-                && self.plugins.is_installed("markdown");
-            let h = if md {
-                self.md_editor_scroll.clone()
-            } else {
-                self.file_scroll.clone()
-            };
-            self.file_editor.update(cx, |e, _| e.set_scroll_handle(h));
-        }
-        self.selected_path = Some(rel.clone());
-        if self.open_tabs.contains(&rel) {
-            // Already open. A permanent open promotes it out of the preview slot; a preview
-            // open of an already-permanent tab leaves its permanence alone (just activates).
-            if !preview && self.preview_tab.as_ref() == Some(&rel) {
-                self.preview_tab = None;
-            }
-        } else if preview {
-            // Reuse the single preview slot: replace its path in place if one exists, else
-            // append. The replaced file's tab vanishes — exactly one temporary tab at a time.
-            match self
-                .preview_tab
-                .take()
-                .and_then(|prev| self.open_tabs.iter().position(|t| t == &prev))
-            {
-                Some(i) => self.open_tabs[i] = rel.clone(),
-                None => self.open_tabs.push(rel.clone()),
-            }
-            self.preview_tab = Some(rel.clone());
-        } else {
-            self.open_tabs.push(rel.clone());
-        }
-        self.open_path = Some(rel);
-        // Scroll the (possibly off-screen) active tab into view on next paint.
-        if let Some(i) = self
-            .open_path
-            .as_ref()
-            .and_then(|p| self.open_tabs.iter().position(|t| t == p))
-        {
-            self.tab_scroll.scroll_to_item(i);
-        }
-        self.load_font_preview(cx);
-    }
 
-    /// If the open file is a font and the "font" plugin is installed, parse its family name
-    /// and register it with the text system so the preview pane can render it. Otherwise
-    /// clears the cached preview. Cheap + idempotent (skips re-registering the same path).
-    pub(crate) fn load_font_preview(&mut self, cx: &mut Context<Self>) {
-        let Some(rel) = self.open_path.clone().filter(|p| is_font_file(p)) else {
-            self.font_preview = None;
-            return;
-        };
-        if !self.plugins.is_installed("font") {
-            self.font_preview = None;
-            return;
-        }
-        if self.font_preview.as_ref().is_some_and(|(p, _)| *p == rel) {
-            return; // already loaded
-        }
-        let abs = self
-            .repo_root
-            .as_ref()
-            .map(|r| r.join(&rel))
-            .unwrap_or_else(|| rel.clone());
-        let Ok(bytes) = std::fs::read(&abs) else {
-            self.font_preview = None;
-            return;
-        };
-        let Some(family) = font_family_name(&bytes) else {
-            self.font_preview = None;
-            return;
-        };
-        // Register the face so `.font_family(family)` resolves to it (idempotent in gpui).
-        let _ = cx
-            .text_system()
-            .add_fonts(vec![std::borrow::Cow::Owned(bytes)]);
-        self.font_preview = Some((rel, SharedString::from(family)));
-    }
 
-    /// Close the tab at `idx`. If it was active, fall to its right neighbour (else left,
-    /// else nothing open).
-    pub(crate) fn close_tab(&mut self, idx: usize, cx: &mut Context<Self>) {
-        if idx >= self.open_tabs.len() {
-            return;
-        }
-        let closing = self.open_tabs.remove(idx);
-        if self.preview_tab.as_ref() == Some(&closing) {
-            self.preview_tab = None;
-        }
-        if self.open_path.as_ref() == Some(&closing) {
-            let next = self
-                .open_tabs
-                .get(idx)
-                .or_else(|| self.open_tabs.get(idx.saturating_sub(1)))
-                .cloned();
-            match next {
-                Some(p) => self.open_file(p, cx),
-                None => self.clear_open(cx),
-            }
-        }
-        self.close_menu(cx);
-    }
 
-    /// Close every tab except the one at `idx`, and make it active.
-    pub(crate) fn close_other_tabs(&mut self, idx: usize, cx: &mut Context<Self>) {
-        let Some(keep) = self.open_tabs.get(idx).cloned() else {
-            return;
-        };
-        self.open_tabs = vec![keep.clone()];
-        // Drop a stale preview pointer if its tab was among those closed.
-        self.preview_tab = self.preview_tab.take().filter(|p| p == &keep);
-        self.open_file(keep, cx);
-        self.close_menu(cx);
-    }
 
-    /// Close all tabs to the right of `idx`. If the active tab was among them, activate `idx`.
-    pub(crate) fn close_tabs_right(&mut self, idx: usize, cx: &mut Context<Self>) {
-        if idx + 1 >= self.open_tabs.len() {
-            self.close_menu(cx);
-            return;
-        }
-        let active_removed = self
-            .open_path
-            .as_ref()
-            .and_then(|p| self.open_tabs.iter().position(|t| t == p))
-            .is_some_and(|pos| pos > idx);
-        self.open_tabs.truncate(idx + 1);
-        // Drop a stale preview pointer if its tab was truncated away.
-        if self
-            .preview_tab
-            .as_ref()
-            .is_some_and(|p| !self.open_tabs.contains(p))
-        {
-            self.preview_tab = None;
-        }
-        if active_removed {
-            if let Some(p) = self.open_tabs.get(idx).cloned() {
-                self.open_file(p, cx);
-            }
-        }
-        self.close_menu(cx);
-    }
 
 
 
@@ -804,7 +575,7 @@ impl Kyde {
 
 
     /// Reset the editor to nothing-open.
-    fn clear_open(&mut self, cx: &mut Context<Self>) {
+    pub(crate) fn clear_open(&mut self, cx: &mut Context<Self>) {
         self.open_path = None;
         self.preview_tab = None;
         self.file_editor.update(cx, |e, cx| {
@@ -979,87 +750,8 @@ impl Kyde {
 
 #[cfg(feature = "terminal")]
 impl Kyde {
-    /// Toggle the bottom terminal panel. Opening it spawns the first tab (lazily, so a
-    /// build that never opens a terminal pays no PTY cost) and focuses it.
-    pub(crate) fn act_toggle_terminal(
-        &mut self,
-        _: &crate::ToggleTerminal,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.term_open = !self.term_open;
-        if self.term_open {
-            if self.term_tabs.is_empty() {
-                self.new_terminal_tab(cx);
-            }
-            // Open in the user's persisted maximized state.
-            self.term_maximized = crate::load_ui_bool("terminal_maximized", false);
-            self.focus_active_terminal(window, cx);
-        } else {
-            self.term_maximized = false;
-        }
-        cx.notify();
-    }
 
-    /// ⌘T while the terminal is focused: open a fresh tab (panel already open) and focus it.
-    pub(crate) fn act_new_terminal_tab(
-        &mut self,
-        _: &crate::NewTerminalTab,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        if !self.term_open {
-            self.term_open = true;
-        }
-        self.new_terminal_tab(cx);
-        self.focus_active_terminal(window, cx);
-        cx.notify();
-    }
 
-    /// Spawn a new terminal tab rooted at the open project (or `$HOME`) and make it active.
-    pub(crate) fn new_terminal_tab(&mut self, cx: &mut Context<Self>) {
-        let cwd = self.repo_root.clone();
-        let view = cx.new(|cx| crate::terminal::TerminalView::new(cwd, cx));
-        // Repaint on title change; close the tab when its context menu's Close is hit.
-        cx.subscribe(&view, |this, v, ev, cx| match ev {
-            crate::terminal::TerminalEvent::TitleChanged => cx.notify(),
-            crate::terminal::TerminalEvent::CloseRequested => {
-                if let Some(idx) = this.term_tabs.iter().position(|t| t == &v) {
-                    this.close_terminal_tab(idx, cx);
-                }
-            }
-        })
-        .detach();
-        self.term_tabs.push(view);
-        self.term_active = self.term_tabs.len() - 1;
-        cx.notify();
-    }
 
-    /// Close a terminal tab; closing the last one hides the panel.
-    pub(crate) fn close_terminal_tab(&mut self, idx: usize, cx: &mut Context<Self>) {
-        if idx >= self.term_tabs.len() {
-            return;
-        }
-        self.term_tabs.remove(idx);
-        if self.term_tabs.is_empty() {
-            self.term_open = false;
-            self.term_maximized = false;
-            self.term_active = 0;
-        } else if self.term_active >= self.term_tabs.len() {
-            self.term_active = self.term_tabs.len() - 1;
-        }
-        cx.notify();
-    }
 
-    /// Move focus to the active terminal tab's widget. Focus now AND next frame via
-    /// `window.defer`: on first open the tab was just spawned this frame, so its
-    /// `TerminalElement` isn't in the window tree yet and an immediate-only focus
-    /// wouldn't stick (same gotcha as the finder/branch-popup focus).
-    pub(crate) fn focus_active_terminal(&self, window: &mut Window, cx: &mut Context<Self>) {
-        if let Some(view) = self.term_tabs.get(self.term_active) {
-            let handle = view.read(cx).handle();
-            window.focus(&handle);
-            window.defer(cx, move |window, _cx| window.focus(&handle));
-        }
-    }
 }
