+++ src/render.rs
@@ -419,640 +419,13 @@ impl Kyde {
 
 
 
-    /// Shared tree row used by Browse, Commit, and Rollback — so items look identical
-    /// everywhere. `checkbox: None` = no checkbox; `Some(checked)` shows one. The three
-    /// closures wire per-site behavior (activate / toggle-check / context-menu).
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn tree_row(
-        &self,
-        cx: &mut Context<Self>,
-        path: &std::path::Path,
-        is_dir: bool,
-        expanded: bool,
-        depth: usize,
-        selected: bool,
-        name: SharedString,
-        name_color: gpui::Rgba,
-        checkbox: Option<bool>,
-        on_activate: impl Fn(&mut Self, &gpui::MouseDownEvent, &mut Window, &mut Context<Self>)
-            + 'static,
-        on_check: impl Fn(&mut Self, &mut Context<Self>) + 'static,
-        on_context: impl Fn(&mut Self, gpui::Point<Pixels>, &mut Context<Self>) + 'static,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let indent = px(8.0 + depth as f32 * 14.0);
 
-        // Caret column (chevron for dirs, empty spacer for files) — fixed width so the
-        // checkbox/badge/name columns align across rows.
-        let caret = div()
-            .w(px(16.0))
-            .flex_none()
-            .text_color(t.line_number)
-            // Larger than the tree's text size for a chunkier chevron.
-            .text_size(px(theme::get().ui_font_size + 5.0))
-            .when(is_dir, |d| d.child(if expanded { "▾" } else { "▸" }));
 
-        // A real drawn checkbox (rounded square; filled with a check svg when ticked),
-        // placed AFTER the caret. Its own click toggles, without firing the row.
-        let checkbox_el = checkbox.map(|checked| {
-            let mut b = div()
-                .flex_none()
-                .size(px(15.0))
-                .mr(px(6.0))
-                .rounded_sm()
-                .border_1()
-                .flex()
-                .items_center()
-                .justify_center();
-            b = if checked {
-                b.bg(t.primary).border_color(t.primary).child(
-                    svg()
-                        .path("icons/check.svg")
-                        .size(px(11.0))
-                        .text_color(t.primary_text),
-                )
-            } else {
-                b.border_color(t.line_number)
-            };
-            b.on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, _e, _w, cx| {
-                    cx.stop_propagation();
-                    on_check(this, cx);
-                }),
-            )
-        });
 
-        let badge = div()
-            .w(px(22.0))
-            .flex_none()
-            .mr(px(4.0))
-            .overflow_hidden()
-            .flex()
-            .items_center()
-            .justify_end()
-            .child(if is_dir {
-                svg()
-                    .path("icons/folder.svg")
-                    .size(px(16.0))
-                    .text_color(gpui::rgb(0x9AA0A6))
-                    .into_any_element()
-            } else {
-                badge_inner(file_badge(path), 2.0)
-            });
-
-        let content = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .flex_1()
-            .min_w_0()
-            .pl(indent)
-            .child(caret)
-            .when_some(checkbox_el, |d, cb| d.child(cb))
-            .child(badge)
-            .child(
-                div()
-                    .flex_1()
-                    .min_w_0()
-                    .truncate()
-                    .text_color(name_color)
-                    .child(name),
-            );
-
-        div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .h(px(38.0))
-            // Set the row's font here (not on the container) so every tree — browse, commit,
-            // rollback — renders identically regardless of which island hosts it.
-            .text_size(px(theme::get().ui_font_size + 3.0))
-            // Never let the flex column shrink a row: with many rows (root expanded) the
-            // shrink would squash every row below its height, making a collapsed root's rows
-            // look taller. `flex_none` keeps them a fixed height and lets the list scroll.
-            .flex_none()
-            .mx(px(6.0))
-            .pr_1()
-            .rounded_md()
-            .cursor_pointer()
-            .when(selected, |d| d.bg(t.selected_bg))
-            // No hover tint on the active row (it would override its selected colour).
-            .when(!selected && !self.dragging(Divider::Tree), |d| {
-                d.hover(|d| d.bg(t.bg_mid))
-            })
-            .child(content)
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, e: &gpui::MouseDownEvent, w, cx| {
-                    on_activate(this, e, w, cx)
-                }),
-            )
-            .on_mouse_down(
-                MouseButton::Right,
-                cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                    on_context(this, e.position, cx)
-                }),
-            )
-            .into_any_element()
-    }
-
-    /// The flattened Browse tree rows: the repo root, its expanded descendants (only when
-    /// the root is open), then a virtual "Scratches" folder + its files at the bottom.
-    /// Depths are offset by one so everything nests under the root row. The "Scratches"
-    /// sentinel path can't collide with a real file.
-    fn browse_visible_rows(&self) -> Vec<tree::Row> {
-        let mut visible = vec![tree::Row {
-            path: PathBuf::new(),
-            is_dir: true,
-            depth: 0,
-        }];
-        if self.expanded.contains(&PathBuf::new()) {
-            for mut r in self.file_tree.visible(&self.expanded) {
-                r.depth += 1;
-                visible.push(r);
-            }
-        }
-        let scratch_group = scratch_group_path();
-        if !self.scratches.is_empty() {
-            visible.push(tree::Row {
-                path: scratch_group.clone(),
-                is_dir: true,
-                depth: 0,
-            });
-            if self.expanded.contains(&scratch_group) {
-                for s in &self.scratches {
-                    visible.push(tree::Row {
-                        path: s.clone(),
-                        is_dir: false,
-                        depth: 1,
-                    });
-                }
-            }
-        }
-        visible
-    }
-
-    /// The Browse right-hand pane: the no-file shortcuts screen when no tab is open,
-    /// otherwise the tab bar + optional install/find banners + the editor area (inline
-    /// image preview, font view, code editor, or the editor | markdown-preview split).
-    fn render_editor_pane(
-        &mut self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let right = if self.open_tabs.is_empty() {
-            div()
-                .flex()
-                .flex_col()
-                .flex_1()
-                .child(self.render_no_file(ui, cx))
-        } else {
-            let image = self.open_path.clone().filter(|p| is_image(p));
-            let editor_area = if let Some(rel) = &image {
-                // Inline image preview, centered, scaled to fit the pane.
-                let abs = self
-                    .repo_root
-                    .as_ref()
-                    .map(|r| r.join(rel))
-                    .unwrap_or_else(|| rel.clone());
-                div()
-                    .id("image-scroll")
-                    .overflow_scroll()
-                    .flex_1()
-                    .flex()
-                    .items_center()
-                    .justify_center()
-                    .p_4()
-                    .child(img(abs).max_w_full().max_h_full())
-                    .into_any_element()
-            } else if self.open_path.as_ref().is_some_and(|p| is_font_file(p)) {
-                self.render_font_view(cx).into_any_element()
-            } else {
-                // Outer flex item is `flex_1 min_w_0 overflow_hidden` — it shrinks to its flex
-                // share regardless of the (content-wide) editor element inside. The INNER
-                // `overflow_scroll` div holds the wide editor and scrolls it. Decoupling the
-                // two is what lets a side-by-side pane (markdown preview) keep its 50%: a
-                // content-wide element directly on the flex item refuses to shrink below it.
-                // Single scroll div holding the content-wide editor element — the explicit width
-                // (set in `editor_with_scrollbars`) goes on THIS div, the same one that has
-                // `overflow_scroll` and the editor as a direct child. (An extra wrapper level
-                // around it lets the content's min-width win and the pane refuses to shrink.)
-                // The editor sits inside an explicit-width wrapper (its content width) so long
-                // lines overflow this `overflow_scroll` viewport and the horizontal scrollbar
-                // appears; `min_h_full` keeps the click target covering short files.
-                let content_w = self.file_editor.read(cx).content_width();
-                let editor_pane = div()
-                    .id("editor-scroll")
-                    .overflow_scroll()
-                    .track_scroll(&self.file_scroll)
-                    // A click in the empty area below a short file forwards to the editor,
-                    // so the caret jumps to the last line (the editor consumes its own text
-                    // clicks via `stop_propagation`, so this only fires for the gap).
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                            let fe = this.file_editor.clone();
-                            fe.update(cx, |ed, cx| {
-                                ed.click_at(e.position, e.modifiers.shift, window, cx)
-                            });
-                            cx.stop_propagation();
-                        }),
-                    )
-                    .child(
-                        div()
-                            .flex_none()
-                            .w(px(content_w))
-                            .min_h_full()
-                            .child(self.file_editor.clone()),
-                    );
-                let island_w = self.editor_island_w(window);
-                let editor = self.with_scrollbars(
-                    editor_pane,
-                    self.file_scroll.clone(),
-                    island_w,
-                    SbView::Editor,
-                    window,
-                    cx,
-                );
-                // Markdown + the markdown plugin installed → editor | live preview side by side.
-                let md = self
-                    .open_path
-                    .as_ref()
-                    .is_some_and(|p| matches!(Lang::from_path(p), Lang::Markdown))
-                    && self.plugins.is_installed("markdown");
-                if md {
-                    let text = self.file_editor.read(cx).text().to_string();
-                    // Two side-by-side panes, each with its own reusable scrollbars: the code
-                    // editor (left, `md_editor_w` wide, drag-resizable divider) and the rendered
-                    // preview (right, the remaining width). Each is wrapped in an explicit-width
-                    // container so the split keeps its proportions.
-                    let preview_w = (island_w - self.md_editor_w - theme::FRAME_GAP).max(120.0);
-                    let code_pane = div()
-                        .id("md-editor-scroll")
-                        .overflow_scroll()
-                        .track_scroll(&self.md_editor_scroll)
-                        .child(
-                            div()
-                                .flex_none()
-                                .w(px(content_w))
-                                .min_h_full()
-                                .child(self.file_editor.clone()),
-                        );
-                    // `flex flex_col` is load-bearing: the `with_scrollbars` output is `flex_1`,
-                    // which only resolves to the pane height when its parent is a flex column.
-                    // Without it the pane grew to content height → never overflowed → no thumb.
-                    let left = div()
-                        .flex()
-                        .flex_col()
-                        .flex_none()
-                        .w(px(self.md_editor_w))
-                        .h_full()
-                        .min_h_0()
-                        .child(self.with_scrollbars(
-                            code_pane,
-                            self.md_editor_scroll.clone(),
-                            self.md_editor_w,
-                            SbView::MdEditor,
-                            window,
-                            cx,
-                        ));
-                    // Persistent selectable rendered-markdown view (keeps its text selection
-                    // across frames); re-parses only when the source actually changes.
-                    // Directory of the open markdown file, for resolving relative image paths.
-                    let base_dir = self.open_path.as_ref().and_then(|p| {
-                        self.repo_root
-                            .as_ref()
-                            .map(|r| r.join(p))
-                            .and_then(|abs| abs.parent().map(|d| d.to_path_buf()))
-                    });
-                    let mv = self
-                        .md_view
-                        .get_or_insert_with(|| {
-                            cx.new(|cx| mdview::MarkdownView::new(&text, base_dir.clone(), cx))
-                        })
-                        .clone();
-                    mv.update(cx, |v, cx| v.set_text(&text, base_dir.clone(), cx));
-                    let preview_pane = div()
-                        .id("md-preview-scroll")
-                        .overflow_y_scroll()
-                        .track_scroll(&self.md_preview_scroll)
-                        .child(mv);
-                    let right = div()
-                        .flex()
-                        .flex_col()
-                        .flex_none()
-                        .w(px(preview_w))
-                        .h_full()
-                        .min_h_0()
-                        .child(self.with_scrollbars(
-                            preview_pane,
-                            self.md_preview_scroll.clone(),
-                            preview_w,
-                            SbView::MdPreview,
-                            window,
-                            cx,
-                        ));
-                    div()
-                        .id("md-split")
-                        .flex_1()
-                        .flex()
-                        .flex_row()
-                        .min_h_0()
-                        .child(left)
-                        .child(
-                            div()
-                                .id("md-divider")
-                                .w(px(theme::FRAME_GAP))
-                                .flex_none()
-                                .h_full()
-                                .cursor_col_resize()
-                                .bg(theme::get().divider)
-                                .on_mouse_down(
-                                    MouseButton::Left,
-                                    cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                                        this.start_divider_drag(
-                                            Divider::MdSplit,
-                                            e.position,
-                                            window,
-                                        );
-                                        cx.notify();
-                                    }),
-                                ),
-                        )
-                        .child(right)
-                        .into_any_element()
-                } else {
-                    editor
-                }
-            };
-            // Right-click in the editor pane shows git commands only (Commit / Rollback /
-            // Push) for the open file — no file-management items.
-            let menu_path = self.open_path.clone();
-            let editor_area = div()
-                .flex()
-                .flex_col()
-                .flex_1()
-                .min_w_0()
-                .min_h_0()
-                .child(editor_area)
-                .on_mouse_down(
-                    MouseButton::Right,
-                    cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                        if let Some(p) = menu_path.clone() {
-                            this.open_menu(e.position, MenuTarget::EditorGit(p), cx);
-                        }
-                    }),
-                );
-            let mut r = div()
-                .flex()
-                .flex_col()
-                .flex_1()
-                .min_h_0()
-                // Without min_w_0 the wide editor content (long code lines) lets this column
-                // grow past the island, and the island's overflow_hidden then clips its right
-                // edge — which is exactly where the tab-bar's trailing controls live, so the
-                // `▾` overflow button vanished once tabs/content got wide.
-                .min_w_0()
-                .child(self.render_tab_bar(ui, fs, cx));
-            // Install banner only applies to text/code files, not image previews.
-            if image.is_none() {
-                if let Some(pack) = self.pending_pack() {
-                    r = r.child(self.render_install_banner(pack, ui, fs, cx));
-                }
-                if self.find_open {
-                    r = r.child(self.render_find_bar(ui, cx));
-                }
-            }
-            r.child(editor_area)
-        };
-        right.into_any_element()
-    }
-
-    fn render_browse(
-        &mut self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        // Row-hover suppression during a tree resize now lives in `tree_row` (it reads
-        // `dragging(Divider::Tree)`), so the cursor sweeping over rows mid-drag doesn't flicker.
-        // Show the repo root as the top row; everything else nests one level under it,
-        // and collapsing the root hides the whole tree.
-        let root_name = self
-            .repo_root
-            .as_ref()
-            .and_then(|p| p.file_name())
-            .map(|s| s.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "/".to_string());
-        let visible = self.browse_visible_rows();
-        let scratch_group = scratch_group_path();
-        // O(1) status lookup per row (was `self.files.iter().find()` — O(changed) per row,
-        // i.e. O(rows×changed) for the whole tree on every frame).
-        let status_by_path: std::collections::HashMap<&PathBuf, FileStatus> =
-            self.files.iter().map(|f| (&f.path, f.status)).collect();
-        let rows: Vec<gpui::AnyElement> = visible
-            .into_iter()
-            .map(|r| {
-                let is_root = r.path.as_os_str().is_empty();
-                let name: SharedString = if is_root {
-                    root_name.clone().into()
-                } else if r.path == scratch_group {
-                    "Scratches".into()
-                } else {
-                    r.path
-                        .file_name()
-                        .map(|n| n.to_string_lossy().into_owned())
-                        .unwrap_or_default()
-                        .into()
-                };
-                let expanded = self.expanded.contains(&r.path);
-                let selected = self.selected_path.as_ref() == Some(&r.path);
-                let (p_act, p_ctx) = (r.path.clone(), r.path.clone());
-                let is_dir = r.is_dir;
-                // Color changed files by their git status (modified/added/…), matching the
-                // commit view; unchanged files use the normal text color.
-                let name_color = (!is_dir)
-                    .then(|| status_by_path.get(&r.path))
-                    .flatten()
-                    .map(|&s| status_color(s))
-                    .unwrap_or(theme::get().text);
-                self.tree_row(
-                    cx,
-                    &r.path,
-                    is_dir,
-                    expanded,
-                    r.depth,
-                    selected,
-                    name,
-                    name_color,
-                    None,
-                    move |this, e, window, cx| {
-                        // Folders toggle expansion. Files (VS Code-style): single click opens in
-                        // the temporary *preview* tab (reused by the next single-click), double
-                        // click opens a permanent tab.
-                        this.selected_path = Some(p_act.clone());
-                        // Focus the app root so the "Kyde"-context Backspace (delete) binding
-                        // is live on the selected row. (open_file/preview_file re-focus the
-                        // editor below, which is what we want there.)
-                        window.focus(&this.focus_handle);
-                        if is_dir {
-                            this.toggle_dir(p_act.clone(), cx);
-                        } else if e.click_count >= 2 {
-                            this.open_file(p_act.clone(), cx);
-                        } else {
-                            this.preview_file(p_act.clone(), cx);
-                        }
-                        cx.notify();
-                    },
-                    |_this, _cx| {},
-                    move |this, pos, cx| {
-                        this.open_menu(pos, MenuTarget::BrowseFile(p_ctx.clone(), is_dir), cx);
-                    },
-                )
-            })
-            .collect();
-        let t = theme::get();
-        // Header strip with a `−` minimize button at the top-right.
-        // Collapse button — absolutely positioned in the tree's top-right corner so it floats
-        // over the rows instead of consuming a whole header line.
-        let tree_minimize = div()
-            .id("tree-minimize")
-            .absolute()
-            .top(px(4.0))
-            // Clear the 12px scrollbar gutter on the right so the bar never sits under it.
-            .right(px(8.0))
-            .flex()
-            .items_center()
-            .justify_center()
-            .size(px(22.0))
-            .rounded_md()
-            .text_size(px(16.0))
-            .text_color(t.line_number)
-            .bg(t.panel_bg)
-            .hover(|s| s.bg(t.bg_light).text_color(t.text))
-            .cursor_pointer()
-            .child("−")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.tree_collapsed = true;
-                    cx.notify();
-                }),
-            );
-        let tree = div()
-            .flex()
-            .flex_col()
-            .relative()
-            .w(px(self.tree_width))
-            .flex_none()
-            .h_full()
-            .py_1()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .text_color(t.text)
-            .font_family(ui)
-            // A touch larger than the editor/code size for readability.
-            .text_size(px(theme::get().ui_font_size + 3.0))
-            .child({
-                let rows_pane = div()
-                    .id("browse-tree")
-                    .overflow_y_scroll()
-                    .track_scroll(&self.tree_scroll)
-                    .flex()
-                    .flex_col()
-                    .children(rows);
-                self.with_scrollbars(
-                    rows_pane,
-                    self.tree_scroll.clone(),
-                    self.tree_width,
-                    SbView::Tree,
-                    window,
-                    cx,
-                )
-            })
-            .child(tree_minimize);
-        // Collapsed: a thin strip with an expand button where the tree was.
-        let collapsed_strip = div()
-            .flex_none()
-            .h_full()
-            .w(px(30.0))
-            .py_1()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .flex()
-            .flex_col()
-            .items_center()
-            .child(
-                div()
-                    .id("tree-expand")
-                    .flex()
-                    .items_center()
-                    .justify_center()
-                    .size(px(24.0))
-                    .rounded_md()
-                    .text_size(px(15.0))
-                    .text_color(t.line_number)
-                    .hover(|s| s.bg(t.bg_light).text_color(t.text))
-                    .cursor_pointer()
-                    .child("»")
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| {
-                            this.tree_collapsed = false;
-                            cx.notify();
-                        }),
-                    ),
-            );
-
-        // The frame-colored gap between the two islands doubles as the resize handle.
-        // No hover/active tint — just the resize cursor.
-        let divider = div()
-            .id("browse-divider")
-            .w(px(theme::FRAME_GAP))
-            .h_full()
-            .flex_none()
-            .cursor_col_resize()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                    this.start_divider_drag(Divider::Tree, e.position, window);
-                    cx.notify();
-                }),
-            );
-
-        // No tabs open → show useful shortcuts instead of the empty editor.
-        let right = self.render_editor_pane(ui, fs, window, cx);
-
-        let editor_island = div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_w_0()
-            .h_full()
-            // `relative` so the floated tab-overflow button anchors to the island's right
-            // edge; `overflow_hidden` keeps it (and the tab strip) clipped to the island.
-            .bg(theme::get().main_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .overflow_hidden()
-            .child(right);
-
-        let mut row = div().flex().flex_row().flex_1();
-        row = if self.tree_collapsed {
-            row.child(collapsed_strip)
-                .child(div().w(px(theme::FRAME_GAP)).flex_none().h_full())
-        } else {
-            row.child(tree).child(divider)
-        };
-        row.child(editor_island).into_any_element()
-    }
 
     /// Width of the editor island = window − rail − body right-pad − (tree + divider, or the
     /// collapsed strip + gap). Used to size the browse editor and the markdown split.
-    fn editor_island_w(&self, window: &Window) -> f32 {
+    pub(crate) fn editor_island_w(&self, window: &Window) -> f32 {
         let win_w = f32::from(window.viewport_size().width);
         let left_chrome = if self.tree_collapsed {
             RAIL_W + theme::FRAME_GAP + 30.0 + theme::FRAME_GAP
@@ -1075,7 +448,7 @@ impl Kyde {
     ///
     /// Thumbs are draggable: `on_mouse_down` arms `sb_drag` with this view's handle, the root
     /// `on_mouse_move` scrolls it.
-    fn with_scrollbars(
+    pub(crate) fn with_scrollbars(
         &mut self,
         content: gpui::Stateful<gpui::Div>,
         scroll: ScrollHandle,
@@ -1219,483 +592,12 @@ impl Kyde {
             .into_any_element()
     }
 
-    fn render_font_view(&self, cx: &mut Context<Self>) -> gpui::Stateful<gpui::Div> {
-        let t = theme::get();
-        let base = div()
-            .id("font-view")
-            .overflow_y_scroll()
-            .flex_1()
-            .flex()
-            .flex_col()
-            .bg(t.main_bg);
-        if !self.plugins.is_installed("font") {
-            return base
-                .items_center()
-                .justify_center()
-                .gap_3()
-                .font_family(theme::font::UI_FAMILY)
-                .child(
-                    div()
-                        .text_color(t.text)
-                        .child("Install Font Preview support to view this file."),
-                )
-                .child(
-                    btn_primary("install-font", "Install Font Preview").on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| this.install_pack("font", cx)),
-                    ),
-                );
-        }
-        let Some((_, family)) = self.font_preview.clone() else {
-            return base
-                .items_center()
-                .justify_center()
-                .font_family(theme::font::UI_FAMILY)
-                .child(
-                    div()
-                        .text_color(t.line_number)
-                        .child("Could not read this font file."),
-                );
-        };
-        let quote = "Me fail English? That's unpossible.";
-        let fam = family.clone();
-        let line = move |sz: f32| {
-            div()
-                .font_family(fam.clone())
-                .text_size(px(sz))
-                .text_color(t.text)
-                .child(quote)
-        };
-        base.p_6()
-            .gap_3()
-            .child(
-                div()
-                    .font_family(theme::font::UI_FAMILY)
-                    .text_color(t.line_number)
-                    .text_size(px(12.0))
-                    .child(family.clone()),
-            )
-            .child(line(13.0))
-            .child(line(18.0))
-            .child(line(24.0))
-            .child(line(32.0))
-            .child(line(48.0))
-            .child(
-                div()
-                    .font_family(family)
-                    .text_size(px(20.0))
-                    .text_color(t.secondary_text)
-                    .child("ABCDEFG abcdefg 0123456789 !?&@#"),
-            )
-    }
-
-    /// Rendered Markdown preview (right pane of the side-by-side markdown view). Live —
-    /// re-renders from the editor's current text each frame.
-    // Superseded by the selectable `mdview::MarkdownView`, but kept (another change adds image
-    // rendering here); left in place rather than deleted.
-    #[allow(dead_code)]
-    fn render_markdown_preview(&self, text: &str) -> gpui::Stateful<gpui::Div> {
-        let t = theme::get();
-        let mk_font = |bold: bool, italic: bool, code: bool| gpui::Font {
-            family: if code {
-                theme::font::FAMILY
-            } else {
-                theme::font::UI_FAMILY
-            }
-            .into(),
-            features: Default::default(),
-            fallbacks: None,
-            weight: if bold {
-                FontWeight::BOLD
-            } else {
-                FontWeight::NORMAL
-            },
-            style: if italic {
-                gpui::FontStyle::Italic
-            } else {
-                gpui::FontStyle::Normal
-            },
-        };
-        let styled = move |spans: &[markdown::Span], color: gpui::Hsla| -> gpui::StyledText {
-            let mut s = String::new();
-            let mut runs = Vec::new();
-            for sp in spans {
-                let c: gpui::Hsla = if sp.code {
-                    gpui::rgb(0xC9CDD6).into()
-                } else {
-                    color
-                };
-                runs.push(gpui::TextRun {
-                    len: sp.text.len(),
-                    font: mk_font(sp.bold, sp.italic, sp.code),
-                    color: c,
-                    background_color: None,
-                    underline: None,
-                    strikethrough: None,
-                });
-                s.push_str(&sp.text);
-            }
-            gpui::StyledText::new(SharedString::from(s)).with_runs(runs)
-        };
-        let blocks = markdown::parse(text).into_iter().map(|b| match b {
-            markdown::Block::Heading(level, spans) => {
-                let size = match level {
-                    1 => 26.0,
-                    2 => 21.0,
-                    3 => 18.0,
-                    _ => 15.0,
-                };
-                div()
-                    .pt_2()
-                    .text_size(px(size))
-                    .child(styled(&spans, t.text.into()))
-                    .into_any_element()
-            }
-            markdown::Block::Paragraph(spans) => div()
-                .text_size(px(14.0))
-                .child(styled(&spans, t.text.into()))
-                .into_any_element(),
-            markdown::Block::Code(code) => div()
-                .font_family(theme::font::FAMILY)
-                .text_size(px(13.0))
-                .text_color(t.text)
-                .bg(t.bg_mid)
-                .rounded_md()
-                .p_3()
-                .child(SharedString::from(code))
-                .into_any_element(),
-            markdown::Block::ListItem(depth, spans) => div()
-                .flex()
-                .flex_row()
-                .pl(px(depth as f32 * 14.0))
-                .text_size(px(14.0))
-                .child(div().pr_2().text_color(t.line_number).child("•"))
-                .child(styled(&spans, t.text.into()))
-                .into_any_element(),
-            markdown::Block::Quote(spans) => div()
-                .border_l_2()
-                .border_color(t.divider)
-                .pl_3()
-                .text_size(px(14.0))
-                .child(styled(&spans, t.secondary_text.into()))
-                .into_any_element(),
-            markdown::Block::Rule => div().h(px(1.0)).bg(t.divider).my_2().into_any_element(),
-            markdown::Block::Image { src, .. } => {
-                // Remote (http/https) → loaded via gpui's asset loader, which needs an
-                // HttpClient wired at startup — only present under the `remote-images`
-                // feature; without it the fetch fails and nothing renders (no crash).
-                // Local → resolved relative to the open file's directory, then the repo
-                // root, and read straight off disk (no deps, always works).
-                let source: gpui::ImageSource =
-                    if src.starts_with("http://") || src.starts_with("https://") {
-                        src.clone().into()
-                    } else {
-                        let rel = src.trim_start_matches("./");
-                        let joined = match self.open_path.as_ref().and_then(|p| p.parent()) {
-                            Some(dir) => dir.join(rel),
-                            None => PathBuf::from(rel),
-                        };
-                        self.repo_root
-                            .as_ref()
-                            .map(|r| r.join(&joined))
-                            .unwrap_or(joined)
-                            .into()
-                    };
-                div()
-                    .py_1()
-                    .child(img(source).max_w_full().max_h(px(480.0)))
-                    .into_any_element()
-            }
-        });
-        div()
-            .id("md-preview")
-            .overflow_y_scroll()
-            .track_scroll(&self.md_preview_scroll)
-            .flex()
-            .flex_col()
-            .gap_2()
-            .p_5()
-            .bg(t.main_bg)
-            .font_family(theme::font::UI_FAMILY)
-            .children(blocks)
-    }
 
 
 
-    fn render_tab_bar(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let dirty = self.file_editor.read(cx).dirty;
-        let tabs = self.open_tabs.iter().enumerate().map(|(i, p)| {
-            let active = self.open_path.as_ref() == Some(p);
-            // The preview (temporary) tab renders in italics, like VS Code.
-            let preview = self.preview_tab.as_ref() == Some(p);
-            let name: SharedString = p
-                .file_name()
-                .map(|n| n.to_string_lossy().into_owned())
-                .unwrap_or_default()
-                .into();
-            // Share the git text-status color (modified/added/…) like the tree rows.
-            let status_col = self
-                .files
-                .iter()
-                .find(|f| &f.path == p)
-                .map(|f| status_color(f.status));
-            let icon = div().flex_none().child(badge_inner(file_badge(p), 0.0));
-            // Active+dirty → a dot in place of the close affordance; otherwise an `×`.
-            let grp = SharedString::from(format!("tabgrp-{i}"));
-            let close = div()
-                .id(SharedString::from(format!("tab-close-{i}")))
-                .flex_none()
-                .w(px(18.0))
-                .flex()
-                .items_center()
-                .justify_center()
-                .rounded_sm()
-                .text_size(px(15.0))
-                .text_color(t.line_number)
-                .hover(|d| d.bg(t.bg_light).text_color(t.text))
-                // Inactive tabs hide the close until the tab is hovered.
-                .when(!active, |d| {
-                    d.opacity(0.0).group_hover(grp.clone(), |s| s.opacity(1.0))
-                })
-                .child(if active && dirty { "●" } else { "×" })
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| {
-                        cx.stop_propagation();
-                        this.close_tab(i, cx);
-                    }),
-                );
-            // Rounded pill per tab: active = accent border + faint accent fill; inactive =
-            // transparent (no bg/border) until hovered. border_1 stays so widths don't shift.
-            div()
-                .id(SharedString::from(format!("tab-{i}")))
-                .group(grp.clone())
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_2()
-                .px_3()
-                .h(px(28.0))
-                .flex_none()
-                .rounded_md()
-                .border_1()
-                .cursor_pointer()
-                // Preview (temporary) tab → the same outline + translucent-fill treatment as an
-                // active tab, but in grey instead of the accent blue, so it reads as "tentative".
-                // A permanent active tab keeps the blue; inactive permanent tabs are bare.
-                .when(preview, |d| {
-                    d.bg(gpui::rgba(0x8A909022))
-                        .border_color(gpui::rgb(0x6B7079))
-                })
-                .when(active && !preview, |d| {
-                    d.bg(gpui::rgba(0x3574F026)).border_color(t.primary)
-                })
-                .when(!active && !preview, |d| {
-                    d.border_color(gpui::rgba(0x00000000))
-                        .hover(|d| d.bg(t.bg_mid))
-                })
-                .text_color(if active { t.text } else { t.line_number })
-                .child(icon)
-                .child(
-                    div()
-                        .min_w_0()
-                        .truncate()
-                        .when_some(status_col, |d, c| d.text_color(c))
-                        .child(name),
-                )
-                .child(close)
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(move |this, _e, _w, cx| {
-                        if let Some(p) = this.open_tabs.get(i).cloned() {
-                            this.open_file(p, cx);
-                            cx.notify();
-                        }
-                    }),
-                )
-                .on_mouse_down(
-                    MouseButton::Right,
-                    cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                        this.open_menu(e.position, MenuTarget::Tab(i), cx);
-                    }),
-                )
-        });
 
-        // The `▾` overflow chooser is NOT rendered here — it's floated on the editor island
-        // (see `render_browse`) so it's pinned to the island's right edge and stays visible
-        // however wide the tab strip grows. We only reserve room for it on the right (`pr`)
-        // so the last tab can scroll out from under it.
-        div()
-            .id("tab-bar")
-            .flex()
-            .flex_row()
-            .items_center()
-            .flex_none()
-            .h(px(38.0))
-            .bg(t.panel_bg)
-            // Match the editor island's top corners (gpui clips rectangular, so the
-            // top strip must round itself or it squares off the island corners).
-            .rounded_t(px(theme::ISLAND_RADIUS))
-            .border_b_1()
-            .border_color(t.divider)
-            .font_family(ui)
-            .text_size(fs)
-            .child(
-                div()
-                    .id("tabs-scroll")
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .px_2()
-                    // Keep the last tab clear of the button's solid right edge; the rest of
-                    // the button is a transparent fade tabs can scroll under.
-                    .pr(px(34.0))
-                    .flex_1()
-                    .min_w_0()
-                    .overflow_x_scroll()
-                    .track_scroll(&self.tab_scroll)
-                    // A plain mouse has only a vertical wheel; map it to horizontal so the
-                    // tab strip scrolls. Native horizontal (trackpad) stays with overflow_x.
-                    .on_scroll_wheel(cx.listener(|this, e: &gpui::ScrollWheelEvent, _w, cx| {
-                        let d = e.delta.pixel_delta(px(18.0));
-                        if d.y.abs() > px(0.0) {
-                            let mut off = this.tab_scroll.offset();
-                            off.x += d.y;
-                            this.tab_scroll.set_offset(off);
-                            cx.notify();
-                        }
-                    }))
-                    .children(tabs),
-            )
-            .into_any_element()
-    }
 
-    /// The `▾` tab-overflow chooser, floated absolutely at the top-right of the editor
-    /// island so it's pinned to the island's right edge and stays visible no matter how
-    /// wide the tab strip grows (the strip can overflow + scroll under it). Click → a
-    /// dropdown listing every open tab. Rendered only when tabs are open.
-    fn render_tab_overflow_button(&self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        // Tabs scroll *under* this; a left→right fade (transparent → panel) dissolves them
-        // into the button instead of a hard bordered box, so it reads as part of the strip.
-        let fade_from = gpui::Rgba {
-            a: 0.0,
-            ..t.panel_bg
-        };
-        // Positioned at the ROOT level (not inside the editor island) so it paints ABOVE the
-        // tab strip's scroll layer — a child of the island was drawn *under* the scrolling
-        // tabs and vanished once they filled the right edge. Coords match the tab bar: y =
-        // titlebar(40) + body top pad(FRAME_GAP); x = body right pad(FRAME_GAP).
-        div()
-            .absolute()
-            .top(px(40.0 + theme::FRAME_GAP))
-            .right(px(theme::FRAME_GAP))
-            // Above the tab bar's 1px bottom border; match its rounded top-right corner.
-            .h(px(37.0))
-            .rounded_tr(px(theme::ISLAND_RADIUS))
-            .flex()
-            .items_center()
-            .justify_end()
-            .w(px(56.0))
-            .pr(px(6.0))
-            .occlude()
-            .bg(gpui::linear_gradient(
-                90.0,
-                gpui::linear_color_stop(fade_from, 0.0),
-                gpui::linear_color_stop(t.panel_bg, 0.55),
-            ))
-            .child(
-                div()
-                    .id("tabs-overflow")
-                    .size(px(24.0))
-                    .flex()
-                    .items_center()
-                    .justify_center()
-                    .rounded_md()
-                    .flex_none()
-                    .cursor_pointer()
-                    .hover(|d| d.bg(t.bg_mid))
-                    .child(
-                        svg()
-                            .path("icons/chevron-down.svg")
-                            .size(px(15.0))
-                            .text_color(t.line_number),
-                    )
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, e: &gpui::MouseDownEvent, _w, cx| {
-                            // Drop the dropdown down-left of the cursor (panel ≥180px wide,
-                            // button hugs the right edge) so it never lands off-screen-right.
-                            let at = gpui::point(
-                                (e.position.x - px(180.0)).max(px(8.0)),
-                                e.position.y + px(8.0),
-                            );
-                            this.open_menu(at, MenuTarget::TabList, cx);
-                        }),
-                    ),
-            )
-            .into_any_element()
-    }
 
-    /// Shown in the editor pane when no file is open: the handful of shortcuts that
-    /// actually get you somewhere (keys reflect the active keymap).
-    fn render_no_file(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let key = |name: &str| {
-            self.keymap
-                .key_for(name)
-                .map(|k| pretty_key(&k))
-                .unwrap_or_default()
-        };
-        let row = |label: &'static str, accel: String| {
-            div()
-                .flex()
-                .flex_row()
-                .items_center()
-                .gap_3()
-                .child(div().text_color(t.text).child(label))
-                .child(
-                    div()
-                        .text_color(t.line_number)
-                        .child(SharedString::from(accel)),
-                )
-        };
-
-        let _ = cx;
-        div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .gap_4()
-            .px_12()
-            .justify_center()
-            // No bg: the rounded editor island behind provides the surface, so the
-            // panel's corners stay rounded (gpui clips rectangular, not rounded).
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size + 2.0))
-            .child(row("Go to File", key("go_to_file")))
-            .child(row("Commit", key("commit")))
-            .child(row("Commit view", key("mode_commit")))
-            .child(row("Keymap / Settings", key("open_keymap")))
-            .child(
-                div()
-                    .text_color(t.line_number)
-                    .child("Select a file from the tree to open it"),
-            )
-            .child(
-                div()
-                    .text_color(t.line_number)
-                    .child("Right-click a file to Commit or Rollback"),
-            )
-            .into_any_element()
-    }
 
 
 
@@ -2015,226 +917,4 @@ impl Kyde {
 
 
 
-    /// Bottom terminal panel: a drag-resize divider, a tab strip (one tab per shell +
-    /// a new-tab button), and the active terminal filling the rest. Inset to align
-    /// under the islands (left of the activity rail).
-    #[cfg(feature = "terminal")]
-    fn render_terminal_panel(
-        &mut self,
-        ui: &'static str,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        // Tab strip: heading + one chip per shell + a new-tab button.
-        let mut strip = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .h(px(34.0))
-            .px_3()
-            .flex_none()
-            .child(
-                div()
-                    .font_family(ui)
-                    .text_size(px(13.0))
-                    .text_color(t.text)
-                    .child("Terminal"),
-            );
-        for (i, view) in self.term_tabs.iter().enumerate() {
-            let active = i == self.term_active;
-            let mut title = view.read(cx).title.clone();
-            if view.read(cx).exited {
-                title.push_str(" (exited)");
-            }
-            strip = strip.child(
-                div()
-                    .id(("term-tab", i))
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_1()
-                    .px_2()
-                    .py_0p5()
-                    .rounded_md()
-                    .when(active, |d| d.bg(t.selected_bg))
-                    .when(!active, |d| d.hover(|d| d.bg(t.bg_mid)))
-                    .cursor_pointer()
-                    .font_family(ui)
-                    .text_size(px(12.0))
-                    .text_color(if active {
-                        t.primary_text
-                    } else {
-                        t.secondary_text
-                    })
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(move |this, _e, window, cx| {
-                            this.term_active = i;
-                            this.focus_active_terminal(window, cx);
-                            cx.notify();
-                        }),
-                    )
-                    .child(title)
-                    .child(
-                        div()
-                            .id(("term-tab-x", i))
-                            .px_1()
-                            .rounded_sm()
-                            .hover(|d| d.bg(t.bg_light))
-                            .child("×")
-                            .on_mouse_down(
-                                MouseButton::Left,
-                                cx.listener(move |this, _e, _w, cx| {
-                                    cx.stop_propagation();
-                                    this.close_terminal_tab(i, cx);
-                                }),
-                            ),
-                    ),
-            );
-        }
-        strip = strip.child(
-            div()
-                .id("term-tab-new")
-                .size(px(22.0))
-                .flex()
-                .items_center()
-                .justify_center()
-                .rounded_md()
-                .hover(|d| d.bg(t.bg_mid))
-                .cursor_pointer()
-                .text_color(t.secondary_text)
-                .tooltip(move |_w, cx| cx.new(|_| Tip("New terminal tab".into())).into())
-                .child("+")
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, window, cx| {
-                        this.new_terminal_tab(cx);
-                        this.focus_active_terminal(window, cx);
-                        cx.notify();
-                    }),
-                ),
-        );
-        // Spacer pushes the minimize + maximize/restore toggles to the right edge of the strip.
-        let maxed = self.term_maximized;
-        strip = strip
-            .child(div().flex_1().min_w_0())
-            // Minimize: hide the panel (reopen via the rail icon or ⌃`). A `−` glyph styled
-            // like the tree / commit-panel collapse buttons.
-            .child(
-                div()
-                    .id("term-minimize")
-                    .size(px(22.0))
-                    .flex()
-                    .items_center()
-                    .justify_center()
-                    .rounded_md()
-                    .text_size(px(16.0))
-                    .text_color(t.line_number)
-                    .hover(|d| d.bg(t.bg_light).text_color(t.text))
-                    .cursor_pointer()
-                    .tooltip(move |_w, cx| cx.new(|_| Tip("Minimize terminal".into())).into())
-                    .child("−")
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| {
-                            this.term_open = false;
-                            this.term_maximized = false;
-                            cx.notify();
-                        }),
-                    ),
-            )
-            .child(
-                div()
-                    .id("term-maximize")
-                    .size(px(22.0))
-                    .flex()
-                    .items_center()
-                    .justify_center()
-                    .rounded_md()
-                    // Same colour + hover as the minimize button next to it.
-                    .text_color(t.line_number)
-                    .hover(|d| d.bg(t.bg_light).text_color(t.text))
-                    .cursor_pointer()
-                    .tooltip(move |_w, cx| {
-                        let label = if maxed {
-                            "Restore terminal"
-                        } else {
-                            "Maximize terminal"
-                        };
-                        cx.new(|_| Tip(label.into())).into()
-                    })
-                    .child(
-                        svg()
-                            .path(if maxed {
-                                "icons/minimize-2.svg"
-                            } else {
-                                "icons/maximize-2.svg"
-                            })
-                            .size(px(15.0))
-                            .text_color(t.line_number),
-                    )
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, window, cx| {
-                            this.term_maximized = !this.term_maximized;
-                            // Persist so the terminal reopens maximized next time.
-                            crate::save_ui_bool("terminal_maximized", this.term_maximized);
-                            this.focus_active_terminal(window, cx);
-                            cx.notify();
-                        }),
-                    ),
-            );
-
-        // The active terminal widget (entity → element).
-        let body = self
-            .term_tabs
-            .get(self.term_active)
-            .map(|v| v.clone().into_any_element())
-            .unwrap_or_else(|| div().into_any_element());
-
-        let island = div()
-            .flex()
-            .flex_col()
-            .size_full()
-            .bg(t.main_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .overflow_hidden()
-            .child(strip)
-            .child(div().flex_1().min_h_0().child(body));
-
-        // A thin top divider whose drag resizes the panel (the shared `Divider::Term` drag).
-        let divider = div()
-            .h(px(6.0))
-            .flex_none()
-            .cursor_row_resize()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                    this.start_divider_drag(Divider::Term, e.position, window);
-                    cx.notify();
-                }),
-            );
-
-        div()
-            .flex()
-            .flex_col()
-            // Maximized → fill the right column; otherwise a fixed, drag-resizable height.
-            .when(self.term_maximized, |d| d.flex_1().min_h_0())
-            .when(!self.term_maximized, |d| {
-                d.flex_none().h(px(self.term_height))
-            })
-            // Inside the right column already (right of the full-height rail) → no left pad;
-            // aligns with the body island above it.
-            .pr(px(theme::FRAME_GAP))
-            // Maximized has no body above it, so it needs the top frame gap the body used to
-            // provide; docked mode gets its top spacing from the resize divider instead.
-            .when(self.term_maximized, |d| d.pt(px(theme::FRAME_GAP)))
-            .pb(px(theme::FRAME_GAP))
-            .bg(t.frame_bg)
-            // No resize divider when maximized (it fills the column).
-            .when(!self.term_maximized, |d| d.child(divider))
-            .child(island)
-            .into_any_element()
-    }
 }
